refactor user test
