+++ .dir-locals.el
@@ -1,4 +1,4 @@
 ((nil . ((projectile-enable-caching . t)
-          (projectile-project-name . "CanchaAPP")
-          (projectile-project-root . "/home/sebas/Projects/cancha-app")
+          (projectile-project-name . "reserva.api")
+          (projectile-project-root . "/home/sebas/Projects/reserva.api")
           )))
